trading_bot/requirements.txt
ccxt
pandas
pyyaml
python-dotenv
TA-Lib
